+++ src/resources/views/pages/home.blade.php
@@ -25,6 +25,7 @@
                         <li>Promotions, Products and Portfolios now allow translation (issue #9).</li>
                         <li>Use Laravel authentication, removed Sentry dependency (issue #81).</li>
                         <li>Allow multiple Category images (issue #79).</li>
+                        <li>Add Post Model and Controller (issue #73).</li>
                         <li>More coming... (developing)</li>
                     </ol>
                     <h4>Bug fixes:</h4>